le ficher .conf ne sera plus partagé, correction d'un warning dans engine.cpp et prends le fichier système pour enet.

+++ Cube/Cube/engine.cpp
@@ -1716,7 +1716,7 @@ void Engine::ManageMenuEnterKeyPress()
 		}
 		else if (m_menu->m_currentMenuItem == MP_SETTINGS)
 			m_menu = new Menu(SM_SETTINGS);
-		else if (m_menu->m_currentMenu == MP_CONTROLS)
+		else if ((int)m_menu->m_currentMenu == (int)MP_CONTROLS)
 			m_menu = new Menu(SM_CONTROLS);
 	}
 	else if (m_menu->m_currentMenu == SM_SETTINGS || m_menu->m_currentMenu == SM_SETTING_SELECTED)
@@ -1796,6 +1796,7 @@ void Engine::ManageMenuEnterKeyPress()
 			{
 				m_settings.m_renderdistance = m_menu->m_settingNewValue;
 				m_settings.Save();
+
 
 				m_world.SetUpdateDistance(m_settings.m_renderdistance);
